wip deterministic password generator toy
We take argon2d(master-password) with a salt of zero as our entropy
seed. Then we use BLAKE2-CTR (not a real thing, but pretty much what
BLAKE2X is doing) to generate a pseudorandom bitstream that can be
sampled to produce passwords.

We try to store the counter where the "node offset" location in the
parameter block would be, to be more like BLAKE2X.

We store an on-disk salted argon2 hash of the master password to try to
detect typos. Since that's the only purpose of this hash, it needn't be
that long - a 1 in 4 billion false positive rate is acceptable. This
hash is salted randomly so that it is not immediately obvious if two
different machines are using the same master password. The argon2 salt
parameter has a minimum size, which is why it uses as much salt as it
does.